requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.5
attrs==21.4.0
azure-core==1.24.1
azure-storage-blob==12.12.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cachetools==5.0.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
dominate==2.6.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.0
Flask==2.1.2
Flask-Compress==1.11
flask-nav==0.6
Flask-SQLAlchemy==2.5.1
fonttools==4.32.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
hdbcli==2.12.22
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mouse==0.7.1
msrest==0.7.1
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
oauthlib==3.2.0
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
plotly==5.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==6.0.0
pycparser==2.21
pycryptodomex==3.14.1
pyDataverse==0.3.1
pydeck==0.7.1
Pygments==2.11.2
PyJWT==2.4.0
Pympler==1.0.1
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
snowflake-connector-python==2.7.8
snowflake-sqlalchemy==1.3.4
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.8.1
streamlit-option-menu==0.3.2
streamlit-tags==1.2.8
tenacity==8.0.1
terminado==0.13.3
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
visitor==0.1.3
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
zipp==3.8.0